link.bzl: Make rpaths refer to the runfiles directory
The rpaths as they were constructed went all the way up to the bazel
cache directory and then used the full path to reference dynamic
libraries.

This is conceptually wrong, since bazel targets only see their
`runfiles` directory at runtime, which is in practice enforced by
`bazel run`, which sandboxes the executed commands (at least on
Linux).

Instead of using the full path from the toplevel cache directory, we
use the `short_path` of the file to point to, from the runfiles
directory.