Add libffi.so.* to the output of ghc’s threaded-rts
`libHSrts_thr-ghc*.*` has a runtime dependency on `libffi.so.*` and a
`LD_RUNPATH` that consists of `$ORIGIN:/path/to/glibc`, meaning it
will look into the directory of `libHSrts_thr-ghc*.*` and search for
`libffi.so.*` there.

Even after extended inquiry it is not clear to me how/whether `ld.so`
follows symlinks when expanding `$ORIGIN` and there doesn’t seem to be
a generally agreed on specification for `LD_RUNPATH` apart from the
`glibc` and `linux` source code. The only thing that always seems to
work is linking the dependencies into the same folder, which is what
this commit does.