Move to Bazel v0.20
Fixes #492.

v0.20 has breaking changes that we can't adapt to without
simultaneously losing compatibility with earlier versions. So the
minimum requirement is now Bazel >= v0.20. This is also true of our
upstream dependencies, so we had to update them as part of this
change.

We're now using the new C++ toolchain API. This in principle should
enable us to shed long standing limitations around linking. But that
will have to be kept as future work.

DON'T MERGE YET: this PR is using a fork of Nixpkgs while Bazel
v0.20.0 gets merged upstream.